Update call to build with current timestamp
When we were generating the build.json file for the insertion of the
data into Benchview we were just using the last source timestamp of the
git repo.  This is causing all of the data to have the same timestamp
and it is not displaying properly on the website.  I have added the
current time as a timestamp and that should fix this issue.