Add new label to stability runs
This change adds stability runs to the new Windows Server 2016 label as
well as increasing the number of iterations to 11 and dropping the
first.  We have also added the same stability prefix that we are using
for regular performance runs to try and remove as much noise as possible
from the benchmark.